# feat(core): faster YAML and JSON parser