### Description

Replace the [yaml](https://github.com/eemeli/yaml) library with [js-yaml](https://github.com/nodeca/js-yaml), which provides at least an order of magnitude performance improvement.
`(2.85MB document, 1.3s => 0.19s)`

We now explicitly provide JSON input support.
We used to use the [yaml](https://github.com/eemeli/yaml) to provide fallback handling for JSON, but it was very slow. Now we've moved to using nodejs' built-in `JSON.parse` function, which uses the simdjson library under the hood. It's two orders of magnitude faster than using yaml's fallback strategy for JSON.
`(2.39MB document, 1.17s => 0.03s)`

To benefit from faster JSON parsing, you need to specify files with a `.json` extension. Files parsed with other extensions still use the fallback strategy of the [js-yaml](https://github.com/nodeca/js-yaml) library.

### Checklist

- [x] I have explained the need for this PR and the problem it solves
- [x] I have explained the changes or the new features added to this PR
- [ ] I have added tests corresponding to this change
- [ ] I have updated the documentation to reflect this change
- [x] I have verified that this change is backward compatible

<!--

Note:

1. Mark the PR as draft until it's ready to be reviewed.
2. Always add/update tests for any changes unless you have a good reason.
3. Always update the documentation to reflect the changes made in the PR.
4. Make a new commit to resolve conversations instead of `push -f`.
5. To resolve merge conflicts, merge master instead of rebasing.
6. Use "request review" to notify the reviewer after making changes.
7. Only a reviewer can mark a conversation as resolved.

-->